Moved the buffer level to a function
